feat: poll empty job queues for a duration before backing off
Problem: when a job queue is empty, a job runner backs off immediately and won't process any more jobs until the next cron schedule.
This is a going to be a problem in a master<>slave situation where a master job runner load jobs into the queue before any processing can start.
The slave job runners should be able to wait for a time window to allow for this possibility